Adds editing frequencies

Adds editing frequencies to an existing SummarizedExperimentobject (created by create_se). Currently A to I editing is supported as well as any custom editing events. The SummarizedExperiment with a new assay for editing requences for each site (edit_freq) and new colData columns with the number of edited sites (n_sites) and the fraction of edits (edit_idx) is returned. If make_plots is set to true and no save directory is specified, a list including the SummarizedExperiment object and plots of the number of sites editied and the fraction of editing events will be returned.

Usage

add_editing_frequencies(
  se_object,
  type = "AI",
  edit_from = NULL,
  edit_to = NULL,
  make_plots = FALSE,
  save_dir = NULL,
  edit_frequency = 0.01,
  min_count = 10,
  ...
)

Arguments

se_object

A SummarizedExperiment object created by create_se

type

OPTIONAL the type of editing event to add. Currently, the options are A to I ("AI"), U to C ("UC"), C to U ("CU"), and G to A ("GA"). "AI" is the default. If you want an editing option outside of these, your own custom can be added by setting this to "none".

edit_from

OPTIONAL if not using a pre-built type, you can specify your own editing. This should be a nucleotide (A, C, G, or T) and should correspond to the nucleotide you expect in the reference. Ex. for A to I editing events, this would be "A". If type is not "AI", both edit from and edit_to must be set.

edit_to

OPTIONAL if not using a pre-built type, you can specify your own editing. This should be a nucleotide (A, C, G, or T) and should correspond to the nucleotide you expect after the editing event. Ex. for A to I editing events, this would be "G". If type is not "AI", both edit from and edit_to must be set.

make_plots

OPTIONAL if plots showing the number and frequency of editing events should be made. Default is False. If True, the colData columns containing the treatment and replicate information should be provided. See make_editing_plots

save_dir

OPTIONAL if the plots created should be saved, include a path to the directory. The plots will be saved as "number_of_sites.pdf" and "editing_index.pdf". If no directory is provided, the plots will be returned in a list with the SummarizedExperiment object.

edit_frequency

OPTIONAL the edit frequency used to determine the number of sites. Default is 0.01.

min_count

OPTIONAL the number of reads used to determine the number of edited sites. Default is 10.

...

Options passed to make_editing_plots