Read in bam file as a data.frame

Read in bam file as a data.frame

Usage

bam_to_df(filename, region = NULL, tags = NULL)

Arguments

filename

path to bam file

region

samtools region query string (i.e. chr1:100-1000)

tags

bam tags to return, supplied with type suffix, i.e. "XS:A" for character, "AS:i" for integer, and "CB:Z" for string. Bam entries with missing tags will return empty strings.

Examples

bam_file <- system.file("extdata", "SRR5564277_Aligned.sortedByCoord.out.bam", package = "raer")
res <- bam_to_df(bam_file)
head(res)
#>   chrom start end               name strand
#> 1  SSR3     0 126  SRR5564277.775963      +
#> 2  SSR3     0 129  SRR5564277.936490      +
#> 3  SSR3     0 123 SRR5564277.1037281      +
#> 4  SSR3     0 123 SRR5564277.1252641      +
#> 5  SSR3     0  63 SRR5564277.2172388      +
#> 6  SSR3     0 125 SRR5564277.2528847      +

res <- bam_to_df(bam_file,
                 region = "DHFR:88-100",
                 tags = c("NH:i", "AS:i"))
res
#>    chrom start end               name strand NH  AS
#> 1   DHFR     0 142  SRR5564277.492430      +  1 285
#> 2   DHFR     0 130  SRR5564277.535229      +  1 276
#> 3   DHFR     0 142  SRR5564277.591923      +  1 224
#> 4   DHFR     0 130 SRR5564277.1198151      +  1 238
#> 5   DHFR     0 142 SRR5564277.1574417      +  1 283
#> 6   DHFR     0 102 SRR5564277.1885362      +  1 245
#> 7   DHFR     0 113 SRR5564277.2066569      +  1 252
#> 8   DHFR     0 130 SRR5564277.2141986      +  1 271
#> 9   DHFR     0 141 SRR5564277.2294813      +  1 290
#> 10  DHFR     0 140 SRR5564277.2514222      +  1 287
#> 11  DHFR     0 129 SRR5564277.3137081      +  1 272
#> 12  DHFR     0 130 SRR5564277.3156909      +  1 267
#> 13  DHFR     0 137 SRR5564277.3331619      +  1 286
#> 14  DHFR     0 142 SRR5564277.3493582      +  1 281
#> 15  DHFR    17 115 SRR5564277.1700085      +  1 205
#> 16  DHFR    35 186  SRR5564277.802338      +  1 294
#> 17  DHFR    36 187 SRR5564277.1686385      +  1 292
#> 18  DHFR    36 187 SRR5564277.1800340      +  1 296
#> 19  DHFR    37 188 SRR5564277.2409065      +  1 300
#> 20  DHFR    42 191  SRR5564277.872889      +  4 289
#> 21  DHFR    42 191  SRR5564277.872889      +  4 289
#> 22  DHFR    42 288  SRR5564277.872889      +  4 288
#> 23  DHFR    42 288  SRR5564277.872889      +  4 288
#> 24  DHFR    54 205 SRR5564277.1614785      +  1 300
#> 25  DHFR    54 205 SRR5564277.3615925      +  1 286
#> 26  DHFR    56 207  SRR5564277.298556      +  1 290
#> 27  DHFR    57 208 SRR5564277.1690225      +  1 298
#> 28  DHFR    71 219  SRR5564277.243543      -  1 209
#> 29  DHFR    82 233  SRR5564277.294767      +  1 294
#> 30  DHFR    82 233 SRR5564277.1506211      +  1 294
#> 31  DHFR    82 233 SRR5564277.2447251      +  1 291
#> 32  DHFR    82 233 SRR5564277.2483958      +  1 210
#> 33  DHFR    82 233 SRR5564277.2648394      +  1 292
#> 34  DHFR    82 233 SRR5564277.2690581      +  1 280
#> 35  DHFR    82 233 SRR5564277.3362853      +  1 269
#> 36  DHFR    82 233 SRR5564277.3598070      +  1 295
#> 37  DHFR    82 233 SRR5564277.3645021      +  1 299
#> 38  DHFR    83 234 SRR5564277.3190114      +  1 298
#> 39  DHFR    84 235  SRR5564277.174688      +  1 298
#> 40  DHFR    84 235 SRR5564277.1078393      +  1 294
#> 41  DHFR    84 235 SRR5564277.1520524      +  1 261
#> 42  DHFR    84 235 SRR5564277.1981187      +  1 296
#> 43  DHFR    85 236 SRR5564277.3162784      +  1 277
#> 44  DHFR    86 237 SRR5564277.1512567      +  1 237
#> 45  DHFR    87 238  SRR5564277.369562      +  1 250
#> 46  DHFR    88 239  SRR5564277.419302      +  1 292
#> 47  DHFR    88 239 SRR5564277.1393129      +  1 298
#> 48  DHFR    91 242  SRR5564277.952031      +  1 284
#> 49  DHFR    91 242 SRR5564277.2999363      +  1 298
#> 50  DHFR    93 244 SRR5564277.1640147      +  1 298
#> 51  DHFR    93 244 SRR5564277.1655370      +  1 298
#> 52  DHFR    96 247 SRR5564277.2294813      -  1 290
#> 53  DHFR    98 249 SRR5564277.2066569      -  1 252
#> 54  DHFR    99 249 SRR5564277.2079905      +  1 293