Create RangedSummarizedExperiment

This function will take either a single result from running pileup_res or a list of results (ie for different samples) from running pileup_res and will return a summarized experiment object that contains assays for each column in the pileup_res output. Currently this doesn't add any sample info and can only take a list of results. Potential added functionality would be to allow the user to give an existing SE object as input or to give sample meta data.

Usage

create_se(pileup_res, sample_names = NULL)

Arguments

pileup_res

results from running get_pileup.R, can be one result, a list of results, or a named list of results. If a named list is give, the colData will be named using the names in the list.

sample_names

OPTIONAL A list of names to be added to the SE object. If no sample names are given and pileup_res is not a named list, then default names (ie sample_1, sample_2, ..., sample_n) will be given and a warning will be printed.

Examples

library(SummarizedExperiment)
bamfn <- system.file("extdata", "SRR5564269_Aligned.sortedByCoord.out.md.bam", package = "raer")
bam2fn <- system.file("extdata", "SRR5564277_Aligned.sortedByCoord.out.md.bam", package = "raer")
fafn <- system.file("extdata", "human.fasta", package = "raer")

plps <- get_pileup(c(bamfn, bam2fn), fafn)
names(plps) <- c("sample1", "sample2")
se <- create_se(plps)

assays(se)
#> List of length 9
#> names(9): Ref Var nRef nVar nA nT nC nG nN

colData(se)
#> DataFrame with 2 rows and 1 column
#>              sample
#>         <character>
#> sample1     sample1
#> sample2     sample2

rowRanges(se)
#> GRanges object with 1695 ranges and 0 metadata columns:
#>          seqnames    ranges strand
#>             <Rle> <IRanges>  <Rle>
#>      [1]     SSR3         1      -
#>      [2]     SSR3         2      -
#>      [3]     SSR3         3      -
#>      [4]     SSR3         4      -
#>      [5]     SSR3         5      -
#>      ...      ...       ...    ...
#>   [1691]     DHFR       514      -
#>   [1692]     DHFR       515      -
#>   [1693]     DHFR       516      -
#>   [1694]     DHFR       517      -
#>   [1695]     DHFR       518      -
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths