Identify sites with read coverage

This function will compute read coverage at sites listed in either a GRanges or BED file. Sites passing with at least min_counts read depth will be returned. This function will compute coverage across all alignments and positions, ignoring any cell-barcodes. Generally it is faster to compute coverage across all positions, then filter to those in the GRanges object, rather than only querying the supplied sites.

Usage

filter_by_coverage(bamfile, gr, min_counts, param = NULL, verbose = FALSE, ...)

Arguments

bamfile

Bam file name

gr

GRanges, or file name of BED file, containing site to examine

min_counts

minimum reads counts to require

param

A ScanBamParam object specifying how reads should be filtered. If not supplied the default behavior will ignore alignments marked as secondary, supplementary, or QC-fail.

verbose

If TRUE, print messages

...

Additional arguments to supply to GenomicAlignments::coverage()

Examples

library(rtracklayer)
bam_fn <- raer_example("SRR5564269_Aligned.sortedByCoord.out.md.bam")
sites <- GRanges(
  rep("SSR3", 101),
  IRanges(100:200, width = 1)
)
filter_by_coverage(bam_fn, sites, min_counts = 24)
#> GRanges object with 18 ranges and 1 metadata column:
#>        seqnames    ranges strand |     score
#>           <Rle> <IRanges>  <Rle> | <integer>
#>    [1]     SSR3       107      * |        24
#>    [2]     SSR3       108      * |        24
#>    [3]     SSR3       109      * |        24
#>    [4]     SSR3       110      * |        24
#>    [5]     SSR3       111      * |        24
#>    ...      ...       ...    ... .       ...
#>   [14]     SSR3       120      * |        25
#>   [15]     SSR3       121      * |        25
#>   [16]     SSR3       122      * |        25
#>   [17]     SSR3       123      * |        25
#>   [18]     SSR3       124      * |        24
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths