Generate base counts using pileup

Generate base counts using pileup

Usage

get_pileup(
  bamfiles,
  fafile,
  bedfile = NULL,
  region = NULL,
  chroms = NULL,
  filterParam = FilterParam(),
  outfile_prefix = NULL,
  bedidx = NULL,
  bam_flags = NULL,
  reads = NULL,
  return_data = TRUE,
  BPPARAM = SerialParam(),
  use_index = FALSE,
  bad_reads = NULL,
  umi_tag = NULL,
  verbose = FALSE
)

FilterParam(
  max_depth = 10000,
  min_base_quality = 20L,
  min_mapq = 0L,
  min_nucleotide_depth = 1L,
  library_type = "fr-first-strand",
  only_keep_variants = FALSE,
  trim_5p = 0L,
  trim_3p = 0L,
  indel_dist = 0L,
  splice_dist = 0L,
  homopolymer_len = 0L,
  max_mismatch_type = c(0L, 0L),
  min_read_bqual = c(0, 0),
  min_splice_overhang = 0L,
  min_variant_reads = 0L,
  ignore_query_Ns = FALSE
)

Arguments

bamfiles

paths to 1 or more bam files

fafile

path to fasta file

bedfile

path to bed file with sites or regions to query

region

samtools region query string (i.e. chr1:100-1000)

chroms

chromosomes to process, not to be used with region

filterParam

object of class FilterParam() which specify various filters to apply to reads and sites during pileup.

outfile_prefix

Output prefix for tabix indexed files. If NULL, no files will be produced.

bedidx

A BedFile object, if supplied, pileup will use index generated by indexBed()

bam_flags

bam flags to filter or keep, use Rsamtools::scanBamFlag() to generate

reads

if supplied a fasta file will be written with reads that pass filters and contain variants

return_data

if TRUE, data is returned as a GRanges, if FALSE a character vector of tabix-index files, specified by outfile_prefix, will be returned.

BPPARAM

A BiocParallel class to control parallel execution. Parallel processing occurs per chromosome, so is disabled when run on a single region.

use_index

if TRUE regions supplied in the bedfile will be queried using the bam file index. By default (FALSE) all alignments are queried via streaming and regions in the bedfile are used to select the sites to keep. Using the index can be much faster when querying a small number of sites in large bamfiles. With many sites in the bedfile , or with small bamfiles, this option may be slower than streaming.

bad_reads

a textfile containing read names to exclude from pileup. Readnames should be formated as readid_1 or readid_2 or readid for paired end first read paired-end second read or single end data.

umi_tag

The bam tag containing a UMI sequence. If supplied, multiple reads with the same UMI sequence will only be counted once per position.

verbose

if TRUE, then report progress and warnings.

max_depth

maximum read depth considered at each site

min_base_quality

min base quality score to consider read for pileup

min_mapq

minimum required MAPQ score, can be a vector of values for each bam file

min_nucleotide_depth

min read depth needed to report site

library_type

read orientation, one of fr-first-strand, fr-second-strand, unstranded, and genomic-unstranded. Can supply as a vector to specify for each input bam. Unstranded library type will be reported based on read alignment. genomic-unstranded will report all variants w.r.t the + strand.

only_keep_variants

if TRUE, then only variant sites will be reported (FALSE by default), can be a vector for each input bamfile

trim_5p

Bases to trim from 5' ends of read alignments

trim_3p

Bases to trim from 3' ends of read alignments

indel_dist

Exclude read if site occurs within given distance from indel event in the read

splice_dist

Exclude read if site occurs within given distance from splicing event in the read

homopolymer_len

Exclude site if occurs within homopolymer of given length

max_mismatch_type

Exclude read if it has X different mismatch types (e.g A-to-G, G-to-C, C-to-G, is 3 mismatch types) or Y # of mismatches, must be supplied as a integer vector of length 2. e.g. c(X, Y).

min_read_bqual

Exclude read if more than X percent of the bases have base qualities less than Y. Numeric vector of length 2. e.g. c(0.25, 20)

min_splice_overhang

Exclude read if site is located adjacent to splice site with an overhang of less than given length.

min_variant_reads

Required number of reads containing a variant for a site to be reported. Calculated per bam file, such that if 1 bam file has >= min_variant_reads, then the site will be reported.

ignore_query_Ns

ignored for now

Value

A list containing a GRanges object for each input bam file, or a vector of the output tabixed file names if return_data is FALSE.

Details

Multiple bam files can be processed together, with files being written for each bam file. In this mode the output regions will be consistent across all files. The min_mapq, only_keep_variants, and library_type parameters can be specified for each input files.

See also

Examples

bamfn <- raer_example("SRR5564269_Aligned.sortedByCoord.out.md.bam")
bam2fn <- raer_example("SRR5564277_Aligned.sortedByCoord.out.md.bam")
fafn <- raer_example("human.fasta")

plp <- get_pileup(bamfn, fafn)
plps <- get_pileup(c(bamfn, bam2fn), fafn)
fp <- FilterParam(only_keep_variants = TRUE, min_nucleotide_depth = 55)
get_pileup(bamfn, fafn, filterParam = fp)
#> GRanges object with 7 ranges and 10 metadata columns:
#>       seqnames    ranges strand |         Ref         Var      nRef
#>          <Rle> <IRanges>  <Rle> | <character> <character> <integer>
#>   [1]     DHFR       207      - |           G          GT        55
#>   [2]     DHFR       208      - |           T          TG        55
#>   [3]     DHFR       223      - |           G          GT        56
#>   [4]     DHFR       225      - |           T          TA        55
#>   [5]     DHFR       291      - |           A          AG        54
#>   [6]     DHFR       300      - |           A          AG        56
#>   [7]     DHFR       332      - |           A          AG        55
#>            nVar        nA        nT        nC        nG        nN
#>       <integer> <integer> <integer> <integer> <integer> <integer>
#>   [1]         1         0         1         0        55         0
#>   [2]         1         0        55         0         1         0
#>   [3]         1         0         1         0        56         0
#>   [4]         1         1        55         0         0         0
#>   [5]         2        54         0         0         2         0
#>   [6]         1        56         0         0         1         0
#>   [7]         1        55         0         0         1         0
#>              nX
#>       <integer>
#>   [1]         4
#>   [2]         4
#>   [3]         6
#>   [4]         7
#>   [5]         6
#>   [6]         4
#>   [7]         2
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome