Subset a bam file to contain only certain tags (e.g. cell barcodes)

Subset a bam file to contain only certain tags (e.g. cell barcodes)

Usage

get_tag_bam(bamfile, barcodes, outbam = NULL, pos_sort_output = TRUE, ...)

Arguments

bamfile

input tag indexed bam file

barcodes

character vector of tag values to extract

outbam

optional output bam file name, if not supplied a temporary file will be used

pos_sort_output

if TRUE, sort output bamfile by position and generate a samtools style index

...

Additional arguments passed to Rsamtools::sortBam()

Value

Returns name of output bam file. the output bam file will be positionally sorted and positionally indexed using Rsamtools.

Examples

library(GenomicAlignments)
bam_fn <- system.file("extdata", "5k_neuron_mouse_xf25_1pct_cbsort.bam", package = "raer")
build_tag_index(bam_fn)
#> [1] "/home/runner/work/_temp/Library/raer/extdata/5k_neuron_mouse_xf25_1pct_cbsort.bam.bri"
cbs <- c("AGGATAATCTCAGAAC-1", "TTCGATTTCCCGAGGT-1")
bam_out <- get_tag_bam(bam_fn, barcodes = cbs)
readGAlignments(bam_out, param = ScanBamParam(tag = "CB"))
#> GAlignments object with 7 alignments and 1 metadata column:
#>       seqnames strand       cigar    qwidth     start       end     width
#>          <Rle>  <Rle> <character> <integer> <integer> <integer> <integer>
#>   [1]        2      -       1S90M        91       576       665        90
#>   [2]        2      -         91M        91       595       685        91
#>   [3]        2      -       85M6S        91       618       702        85
#>   [4]        2      -      73M18S        91       626       698        73
#>   [5]        2      -      74M17S        91       629       702        74
#>   [6]        6      - 62M1D16M13S        91       270       348        79
#>   [7]        6      - 48M1I13M29S        91       321       381        61
#>           njunc |                 CB
#>       <integer> |        <character>
#>   [1]         0 | TTCGATTTCCCGAGGT-1
#>   [2]         0 | AGGATAATCTCAGAAC-1
#>   [3]         0 | AGGATAATCTCAGAAC-1
#>   [4]         0 | AGGATAATCTCAGAAC-1
#>   [5]         0 | TTCGATTTCCCGAGGT-1
#>   [6]         0 | TTCGATTTCCCGAGGT-1
#>   [7]         0 | AGGATAATCTCAGAAC-1
#>   -------
#>   seqinfo: 4 sequences from an unspecified genome