A reference class for generating and storing index of BedFile

Class to store information about a bedfile used for indicating regions for pileup

Usage

indexBed(file, ...)

# S4 method for character
indexBed(file)

# S3 method for BedFile
close(con, ...)

Arguments

file

path to bed file

...

present for consistency

con

BedFile class

Value

Return an object of class BedFile which contains a pointer to the bed file index.

Fields

.extptr

an externalptr to c-level samtools bedindex

path

filepath to bed file

open

logical indicating if the index is open

Examples

bed_fn <- system.file("extdata", "regions.bed", package = "raer")
indexBed(bed_fn)
#> Reference class object of class "BedFile"
#> Field ".extptr":
#> Object of class 'externalptr':
#>   x address: 0x55b1393de3d0
#>   R_ExternalPtrAddr(x): 0x55b129d4c480
#>   R_ExternalPtrTag(x): 0x55b1085e41e0
#>   typeof(R_ExternalPtrTag(x)): NULL
#>   R_ExternalPtrProtected(x): 0x55b13ae0f690
#>   typeof(R_ExternalPtrProtected(x)): character
#> Field "path":
#> [1] "/__w/_temp/Library/raer/extdata/regions.bed"
#> Field "open":
#> [1] TRUE
bed_fn <- system.file("extdata", "regions.bed", package = "raer")
bed <- indexBed(bed_fn)
close(bed)