Makes summary plots of editing

Generates plots of the number of sites edited per sample and the percent of editing events per sample. This function is written to be called directly by add_editing_frequencies

Usage

make_editing_plots(
  se_object,
  colors = NULL,
  meta_col = "genotype_treatment",
  replicate = "rep"
)

Arguments

se_object

A SummarizedExperiment object created by create_se

colors

OPTIONAL The colors of the replicates. If no colors are provided, Set1 from `RColorBrewer will be used

meta_col

The column in colData to be used to separate out samples based on the condition. For example, "genotype", "treatment", or "genotype_treatment." Default is "genotype_treatment."

replicate

The column in colData contining information about the replicates. Default is "rep".