Perform differential editing

Uses either edgeR or DESeq2 to perform differential editing analysis. This will work for simple designs that have 1 treatment and 1 control. For more complex designs, we suggest you perform your own.

At the moment, this function will only find editing events specific to the treatment, but it will be pretty straight forward to add other possible return values.

Usage

perform_de(
  deobj,
  type = "edgeR",
  sample_col = "sample",
  condition_col = "condition",
  condition_control = NULL,
  condition_treatment = NULL
)

Arguments

deobj

A SummarizedExperiment object prepared for de by prep_for_de

type

OPTIONAL if edgeR or DESeq should be run. Default is edgeR

sample_col

OPTIONAL the name of the column from colData(deobj) that contains your sample information. Default is sample. If you do not have a column named "sample", you must provide the appropriate sample column

condition_col

OPTIONAL the name of the column from colData(deobj) that contains your treatment information. Default is condition, If you do not have a column named "condition", you must provide the appropriate condition column

condition_control

The name of the control condition. This must be a variable in your condition_col of colData(deobj). No default provided.

condition_treatment

The name of the treatment condition. This must be a variable in your condition_col of colData(deobj).

Value

A named list - de_obj: The edgeR or deseq object used for differential editing analysis - results_full: Unfiltered differenital editing results - sig_results: Filtered differenial editing (FDR < 0.05) - model_matrix: The model matrix used for generating DE results

Examples

example(create_se, echo = FALSE)
se <- calc_edit_frequency(se)
dse <- prep_for_de(se)
res <- perform_de(dse, condition_control = "WT", condition_treatment = "KO")
#> Using classic mode.
res$sig_results[1:5, ]
#>                 logFC   logCPM        LR       PValue         FDR
#> DHFR_361_2 -11.147610 21.96719 13.393676 0.0002524741 0.006816801
#> SSR3_176_2  -8.000087 21.19460 11.166011 0.0008330957 0.011246792
#> SSR3_364_2  -5.600787 19.69710  9.086110 0.0025755817 0.016444635
#> DHFR_260_2  -8.108926 22.26785  8.548454 0.0034581581 0.016444635
#> SSR3_290_2  -4.839287 20.65403  8.454728 0.0036409569 0.016444635