Make summarized experiment object for DE

Generates a SummarizedExperiment object for use with edgeR or DESeq2 will generate a counts assay with a matrix formated with 2 columns per sample

Usage

prep_for_de(
  se,
  type = "AI",
  edit_from = NULL,
  edit_to = NULL,
  min_prop = 0.1,
  max_prop = 0.9,
  min_samples = 3
)

Arguments

se

A SummarizedExperiment object

type

OPTIONAL the type of editing event to add. Currently, only A to I is supported ("AI") which is the default, but your own custom can be added by setting this to "none".

edit_from

OPTIONAL if not using a pre-built type, you can specify your own editing. This should be a nucleotide (A, C, G, or T) and should correspond to the nucleotide you expect in the reference. Ex. for A to I editing events, this would be "A". If type is not "AI", both edit from and edit_to must be set.

edit_to

OPTIONAL if not using a pre-built type, you can specify your own editing. This should be a nucleotide (A, C, G, or T) and should correspond to the nucleotide you expect after the editing event. Ex. for A to I editing events, this would be "G". If type is not "AI", both edit from and edit_to must be set.

min_prop

OPTIONAL the min proporation of reads edited at a site. At least min_samples need to pass this to keep the site. Default is 0.1.

max_prop

OPTIONAL the max proporation of reads edited at a site. At least min_samples need to pass this to keep the site. Default is 0.9.

min_samples

OPTIONAL the minimum number of samples passing the cutoffs to keep a site. Default is 3.

Examples

example(create_se, echo = FALSE)
se <- calc_edit_frequency(se)
dse <- prep_for_de(se)
assay(dse, "counts")
#>             KO1_ref KO2_ref KO3_ref WT1_ref WT2_ref WT3_ref KO1_alt
#> SSR3_176_2       16      16      16       8       8       8       0
#> SSR3_201_2       14      14      14      19      19      19       0
#> SSR3_207_2       14      14      14      14      14      14       0
#> SSR3_257_2       19      19      19      11      11      11       0
#> SSR3_258_2       10      10      10       9       9       9       8
#> SSR3_266_2       17      17      17      12      12      12       0
#> SSR3_279_2       17      17      17       6       6       6       0
#> SSR3_290_2       14      14      14       3       3       3       1
#> SSR3_360_2        6       6       6       3       3       3       0
#> SSR3_364_2        3       3       3       2       2       2       0
#> SSR3_367_2        0       0       0       0       0       0       0
#> SSR3_388_2        0       0       0       0       0       0       0
#> SSR3_502_2        5       5       5       4       4       4       0
#> SPCS3_241_1       2       2       2       3       3       3       0
#> SPCS3_330_1       0       0       0       1       1       1       0
#> SPCS3_338_1       0       0       0       2       2       2       0
#> SPCS3_347_1       4       4       4       5       5       5       0
#> SPCS3_348_1       0       0       0       0       0       0       0
#> DHFR_260_2       60      60      60      21      21      21       0
#> DHFR_263_2       58      58      58      27      27      27       0
#> DHFR_291_2       54      54      54      30      30      30       2
#> DHFR_292_2       53      53      53      26      26      26       0
#> DHFR_309_2       56      56      56      29      29      29       0
#> DHFR_328_2       51      51      51      31      31      31       0
#> DHFR_354_2       47      47      47      28      28      28       0
#> DHFR_355_2       51      51      51      28      28      28       0
#> DHFR_361_2       43      43      43       4       4       4       0
#>             KO2_alt KO3_alt WT1_alt WT2_alt WT3_alt
#> SSR3_176_2        0       0      16      16      16
#> SSR3_201_2        0       0       3       3       3
#> SSR3_207_2        0       0       4       4       4
#> SSR3_257_2        0       0       4       4       4
#> SSR3_258_2        8       8       6       6       6
#> SSR3_266_2        0       0       5       5       5
#> SSR3_279_2        0       0       4       4       4
#> SSR3_290_2        1       1       7       7       7
#> SSR3_360_2        0       0       3       3       3
#> SSR3_364_2        0       0       4       4       4
#> SSR3_367_2        0       0       1       1       1
#> SSR3_388_2        0       0       0       0       0
#> SSR3_502_2        0       0       0       0       0
#> SPCS3_241_1       0       0       1       1       1
#> SPCS3_330_1       0       0       2       2       2
#> SPCS3_338_1       0       0       1       1       1
#> SPCS3_347_1       0       0       0       0       0
#> SPCS3_348_1       0       0       0       0       0
#> DHFR_260_2        0       0      12      12      12
#> DHFR_263_2        0       0       6       6       6
#> DHFR_291_2        2       2       4       4       4
#> DHFR_292_2        0       0      10      10      10
#> DHFR_309_2        0       0       8       8       8
#> DHFR_328_2        0       0       4       4       4
#> DHFR_354_2        0       0       4       4       4
#> DHFR_355_2        0       0       6       6       6
#> DHFR_361_2        0       0      27      27      27
dse
#> class: SummarizedExperiment 
#> dim: 27 12 
#> metadata(0):
#> assays(1): counts
#> rownames(27): SSR3_176_2 SSR3_201_2 ... DHFR_355_2 DHFR_361_2
#> rowData names(0):
#> colnames(12): KO1_ref KO2_ref ... WT2_alt WT3_alt
#> colData names(5): sample condition n_sites edit_idx count