Read pileup, indexed by tabix

This function can read in pileup or site data tables produced by pileup_sites().

Usage

read_pileup(tbx_fn, region = NULL)

Arguments

tbx_fn

filename

region

region to read from file, samtools style region specifiers are supported.

Value

GenomicRanges::GRanges containing positions of editing sites and pileup data.

Examples

bamfn <- system.file("extdata", "SRR5564269_Aligned.sortedByCoord.out.md.bam", package = "raer")
fafn <- system.file("extdata", "human.fasta", package = "raer")
plp_fn <- tempfile()
plp <- pileup_sites(bamfn, fafn, return_data = FALSE, outfile_prefix = plp_fn)

# first table contains site information, second table pileup count
lapply(plp, function(x) head(read_pileup(x)))
#> [[1]]
#> GRanges object with 6 ranges and 4 metadata columns:
#>       seqnames    ranges strand |         REF      rpbz       vdb
#>          <Rle> <IRanges>  <Rle> | <character> <numeric> <numeric>
#>   [1]     SSR3         1      - |           C         0         0
#>   [2]     SSR3         2      - |           A         0         0
#>   [3]     SSR3         3      - |           A         0         0
#>   [4]     SSR3         4      - |           A         0         0
#>   [5]     SSR3         5      - |           G         0         0
#>   [6]     SSR3         6      - |           T         0         0
#>             sor
#>       <numeric>
#>   [1]         0
#>   [2]         0
#>   [3]         0
#>   [4]         0
#>   [5]         0
#>   [6]         0
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths
#> 
#> [[2]]
#> GRanges object with 6 ranges and 10 metadata columns:
#>       seqnames    ranges strand |         REF         ALT      nRef
#>          <Rle> <IRanges>  <Rle> | <character> <character> <integer>
#>   [1]     SSR3         1      - |           C           -        13
#>   [2]     SSR3         2      - |           A           -        14
#>   [3]     SSR3         3      - |           A           -        14
#>   [4]     SSR3         4      - |           A           -        15
#>   [5]     SSR3         5      - |           G           -        15
#>   [6]     SSR3         6      - |           T           -        15
#>            nAlt        nA        nT        nC        nG        nN
#>       <integer> <integer> <integer> <integer> <integer> <integer>
#>   [1]         0         0         0        13         0         0
#>   [2]         0        14         0         0         0         0
#>   [3]         0        14         0         0         0         0
#>   [4]         0        15         0         0         0         0
#>   [5]         0         0         0         0        15         0
#>   [6]         0         0        15         0         0         0
#>              nX
#>       <integer>
#>   [1]         2
#>   [2]         1
#>   [3]         1
#>   [4]         0
#>   [5]         0
#>   [6]         0
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome; no seqlengths
#> 
unlink(c(plp, plp_fn, paste0(plp, ".tbi")))