Calculate editing frequencies

Calculate editing frequencies

Usage

sc_editing(
  bamfile,
  fafile,
  bedfile,
  cell_barcodes,
  min_reads = 25L,
  assay_cols = c("nA", "nG"),
  tag_index_args = list(tag = "CB"),
  ...,
  BPPARAM = SerialParam(),
  verbose = TRUE
)

Arguments

bamfile

BAM file name

fafile

FASTA file name

bedfile

BED file containing editing sites

cell_barcodes

List of character vectors containing cell barcodes to query. See examples for specification.

min_reads

Minimum read counts required to consider a site for editing. This is calculated across all reads in the bamfile prior to running [get_pileup()] per cell to remove low-frequency events. Read alignments are required to be non-duplicate, primary, QC passing, and non-supplemental.

assay_cols

assays to store in returned se. Set to "A" and "G". Note that storing multiple assays can require large amounts of memory.

tag_index_args

arguments pass to build_tag_index()

...

additional arguments passed to [get_pileup()].

BPPARAM

BiocParallel instance. Parallel computation occurs across each entry in the cell_barcodes list

verbose

Display messages

Examples

suppressPackageStartupMessages(library(SummarizedExperiment))
# get vector of cell barcodes in bam file (for use in this example)

bamfn <- raer_example("5k_neuron_mouse_xf25_1pct_cbsort.bam")
idxfn <- build_tag_index(bamfn)
cbs <- show_tag_index(bamfn)$tag
cbs[1:5]
#> [1] "AAACCCAGTCACTTAG-1" "AAACGAACAACTGAAA-1" "AAACGAAGTGAATATG-1"
#> [4] "AAACGCTAGACTTGTC-1" "AAACGCTGTGGATTTC-1"

# process each cell individually
# will be slow with many sites and cells
# bam file will be indexed by build_tag_index() if not already done.
fp <- FilterParam(library_type = "fr-second-strand")
se <- sc_editing(bamfile = bamfn,
  fafile = raer_example("mouse_tiny.fasta"),
  bedfile = raer_example("5k_neuron_sites.bed.gz"),
  min_reads = 0,
  cell_barcodes = cbs[1:15],
  filterParam = fp,
  verbose = FALSE)

# pool cell barcodes across clusters
# pass a named list, with each list entry corresponding to a vector
# of cell barcodes from a cluster

# simulate 5 clusters
cb_lst <- split(cbs, cut(seq_along(cbs), breaks = 5))
names(cb_lst) <- paste0("cluster", 1:5)

se <- sc_editing(bamfile = bamfn,
  fafile = raer_example("mouse_tiny.fasta"),
  bedfile = raer_example("5k_neuron_sites.bed.gz"),
  min_reads = 0,
  cell_barcodes = cb_lst,
  filterParam = fp,
  verbose = FALSE)
assays(se)$nA
#>          cluster1 cluster2 cluster3 cluster4 cluster5
#> 2_548_2         1        2        1        3        2
#> 2_576_2        60       58       56       55       63
#> 2_591_2        94       92       99       96      105
#> 2_593_2        99       95      105      104      108
#> 2_609_2       161      152      162      168      159
#> 2_754_2         0        0        1        0        2
#> 2_755_2         0        0        1        0        2
#> 2_756_2         0        0        1        0        2
#> 2_792_2         0        0        0        0        2
#> 6_271_2         6        9        7        5       10
#> 6_331_2        34       40       34       34       54
#> 6_352_2        32       34       30       29       49
#> 6_361_2        23       26       25       24       39
#> 6_383_2        15       22       15       18       26
#> 6_393_2        13       15        8       10       12
#> 11_366_2        0        0        0        0        0
#> 11_367_2        0        0        0        0        0
assays(se)$nG
#>          cluster1 cluster2 cluster3 cluster4 cluster5
#> 2_548_2         0        0        1        0        0
#> 2_576_2         1        0        0        1        0
#> 2_591_2         0        1        1        1        0
#> 2_593_2         0        1        0        0        0
#> 2_609_2         1        0        0        0        0
#> 2_754_2         0        0        0        0        1
#> 2_755_2         0        0        0        0        1
#> 2_756_2         0        0        0        0        1
#> 2_792_2         0        0        1        0        1
#> 6_271_2         0        1        0        0        0
#> 6_331_2         1        1        0        0        0
#> 6_352_2         0        0        0        2        0
#> 6_361_2         4        2        0        2        2
#> 6_383_2         1        0        0        0        0
#> 6_393_2         1        0        0        0        0
#> 11_366_2        0        1        0        0        0
#> 11_367_2        0        1        0        0        0