Calculate the Adenosine Editing Index (AEI)

The Adenosine Editing Index describes the magnitude of A-to-I editing in a sample. The index is a weighted average of editing events (G bases) observed at A positions. The vast majority A-to-I editing occurs in ALU elements in the human genome, and these regions have a high A-to-I editing signal compared to other regions such as coding exons. This function will perform pileup at specified repeat regions and return a summary AEI metric.

Usage

calc_AEI(
  bam_fn,
  fasta_fn,
  alu_ranges = NULL,
  txdb = NULL,
  snp_db = NULL,
  filterParam = FilterParam(),
  BPPARAM = SerialParam(),
  verbose = FALSE
)

Arguments

bam_fn

bam file

fasta_fn

fasta

alu_ranges

GRanges or the name of a BEDfile with regions to query for calculating the AEI, typically ALU repeats. If a BED file is supplied it will not be filtered by the txdb option.

txdb

A txdb object, if supplied, will be used to subset the alu_ranges to those found overlapping genes. Alternatively a GRanges object with gene coordinates.

snp_db

either a SNPlocs package, GPos, or GRanges object. If supplied, will be used to exclude polymorphic positions prior to calculating the AEI. If calc_AEI() will be used many times, one could save some time by first identifying SNPs that overlap the supplied alu_ranges, and passing these as a GRanges to snp_db rather than supplying all known SNPs. Combined with using a bedfile for alu_ranges can also will save time.

filterParam

object of class FilterParam() which specify various filters to apply to reads and sites during pileup.

BPPARAM

A BiocParallelParam object for specifying parallel options for operating over chromosomes.

verbose

report progress on each chromosome?

Value

A named list with the AEI index computed for all allelic combinations. If correctly computed the signal from the A_G index should be higher than other alleles (T_C), which are most likely derived from noise or polymorphisms.

References

Roth, S.H., Levanon, E.Y. & Eisenberg, E. Genome-wide quantification of ADAR adenosine-to-inosine RNA editing activity. Nat Methods 16, 1131–1138 (2019). https://doi.org/10.1038/s41592-019-0610-9

Examples

suppressPackageStartupMessages(library(Rsamtools))
bamfn <- system.file("extdata", "SRR5564277_Aligned.sortedByCoord.out.md.bam", package = "raer")
fafn <- system.file("extdata", "human.fasta", package = "raer")
dummy_alu_ranges <- scanFaIndex(fafn)
calc_AEI(bamfn, fafn, dummy_alu_ranges)
#> $A_C
#> [1] 0
#> 
#> $A_G
#> [1] 1.879456
#> 
#> $A_T
#> [1] 0.1319261
#> 
#> $C_A
#> [1] 0.2426203
#> 
#> $C_G
#> [1] 0
#> 
#> $C_T
#> [1] 0.2224469
#> 
#> $G_A
#> [1] 0.3120411
#> 
#> $G_C
#> [1] 0
#> 
#> $G_T
#> [1] 0
#> 
#> $T_A
#> [1] 0
#> 
#> $T_C
#> [1] 0
#> 
#> $T_G
#> [1] 0
#>