Apply strand correction using gene annotations

Gene annotations are used to infer the likely strand of editing sites. This function will operate on unstranded datasets which have been processed using "genomic-unstranded" library type which reports variants with respect to the

strand for all sites. The strand of the editing site will be assigned the strand of overlapping features in the genes_gr object. Sites with no-overlap, or overlapping features with conflicting strands (+ and -) will be removed.

Usage

correct_strand(gr, genes_gr)

Arguments

gr

GRanges object containing editing sites processed with "genomic-unstranded" setting

genes_gr

GRanges object containing reference features to annotate the strand of the editing sites.

Examples

suppressPackageStartupMessages(library("GenomicRanges"))
bamfn <- system.file("extdata", "SRR5564269_Aligned.sortedByCoord.out.md.bam", package = "raer")
fafn <- system.file("extdata", "human.fasta", package = "raer")
fp <- FilterParam(library_type = "genomic-unstranded")
plp <- get_pileup(bamfn, fafn, filterParam = fp)

genes <- GRanges(c("DHFR:200-400:+",
  "SPCS3:100-200:-",
  "SSR3:3-10:-",
  "SSR3:6-12:+"))
correct_strand(plp, genes)
#> GRanges object with 307 ranges and 9 metadata columns:
#>         seqnames    ranges strand |         Ref         Var      nRef
#>            <Rle> <IRanges>  <Rle> | <character> <character> <integer>
#>     [1]     SSR3         3      - |           A           -        14
#>     [2]     SSR3         4      - |           A           -        15
#>     [3]     SSR3         5      - |           G           -        15
#>     [4]     SSR3        11      + |           A           -        15
#>     [5]     SSR3        12      + |           T           -        15
#>     ...      ...       ...    ... .         ...         ...       ...
#>   [303]     DHFR       396      + |           T           -        40
#>   [304]     DHFR       397      + |           T           -        43
#>   [305]     DHFR       398      + |           T           -        43
#>   [306]     DHFR       399      + |           C           -        43
#>   [307]     DHFR       400      + |           T           -        45
#>              nVar        nA        nT        nC        nG        nN
#>         <integer> <integer> <integer> <integer> <integer> <integer>
#>     [1]         0        14         0         0         0         0
#>     [2]         0        15         0         0         0         0
#>     [3]         0         0         0         0        15         0
#>     [4]         0        15         0         0         0         0
#>     [5]         0         0        15         0         0         0
#>     ...       ...       ...       ...       ...       ...       ...
#>   [303]         0         0        40         0         0         0
#>   [304]         0         0        43         0         0         0
#>   [305]         0         0        43         0         0         0
#>   [306]         0         0         0        43         0         0
#>   [307]         0         0        45         0         0         0
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome