Create RangedSummarizedExperiment

This function will take either a single result from running get_pileup() or a list of results (ie for different samples) from running pileup_res and will return a summarized experiment object that contains assays for specified columns in the get_pileup() output.

Usage

create_se(
  plps,
  assay_cols = c("Var", "nRef", "nVar", "nA", "nT", "nC", "nG"),
  sample_names = NULL,
  sparse = FALSE,
  fill_na = NULL,
  verbose = FALSE
)

Arguments

plps

results from running get_pileup(), can be one result, a list of results, or a named list of results. If a named list is given, the colData will be named using the names in the list.

assay_cols

character vector of columns to store as assays

sample_names

A list of names to be added to the SE object. If no sample names are given and plps is not a named list, then default names (ie sample_1, sample_2, ..., sample_n) will be given and a warning will be printed.

sparse

If TRUE, numeric matrices will be stored as sparseMatrices. All missing values will be coerced to 0 in the sparseMatrices.

fill_na

Numeric value to replace NAs in numeric matrices Should only be used when plps were computed independently with a min_nucleotide_count = 1, otherwise sites may be set to 0, although they may have coverage > 0 but less than the min_nucleotide_count parameter. Not applied when sparse = TRUE, which coerces missing values to 0.

verbose

print information on progress

Value

RangedSummarizedExperiment object populated with assays for each of the listed assay_cols.

Examples

library(SummarizedExperiment)

bamfn <- raer_example("SRR5564269_Aligned.sortedByCoord.out.md.bam")
bam2fn <- raer_example("SRR5564277_Aligned.sortedByCoord.out.md.bam")
fafn <- raer_example("human.fasta")

bams <- rep(c(bamfn, bam2fn), each = 3)
sample_ids <- paste0(rep(c("KO", "WT"), each = 3), 1:3)
fp <- FilterParam(only_keep_variants = TRUE)
plps <- get_pileup(bams, fafn, filterParam = fp)
names(plps) <- sample_ids
se <- create_se(plps)
se$condition <- substr(se$sample, 1, 2)
assays(se)
#> List of length 7
#> names(7): Var nRef nVar nA nT nC nG

colData(se)
#> DataFrame with 6 rows and 2 columns
#>          sample   condition
#>     <character> <character>
#> KO1         KO1          KO
#> KO2         KO2          KO
#> KO3         KO3          KO
#> WT1         WT1          WT
#> WT2         WT2          WT
#> WT3         WT3          WT

rowRanges(se)
#> GRanges object with 74 ranges and 1 metadata column:
#>              seqnames    ranges strand |         Ref
#>                 <Rle> <IRanges>  <Rle> | <character>
#>   SSR3_102_-     SSR3       102      - |           T
#>   SSR3_125_-     SSR3       125      - |           C
#>   SSR3_156_-     SSR3       156      - |           C
#>   SSR3_176_-     SSR3       176      - |           A
#>   SSR3_198_-     SSR3       198      - |           A
#>          ...      ...       ...    ... .         ...
#>   DHFR_397_-     DHFR       397      - |           A
#>   DHFR_399_-     DHFR       399      - |           G
#>   DHFR_423_-     DHFR       423      - |           T
#>   DHFR_430_-     DHFR       430      - |           A
#>   DHFR_513_-     DHFR       513      - |           T
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome