Generate base counts using pileup

Generate base counts using pileup

Usage

get_pileup(
  bamfile,
  fafile,
  bedfile = NULL,
  region = NULL,
  chrom = NULL,
  start = NULL,
  end = NULL,
  min_reads = c(1L),
  min_base_qual = 20L,
  min_mapq = c(0L),
  max_depth = 10000,
  library_type = "fr-first-strand",
  outfile = NULL,
  bedidx = NULL,
  bam_flags = NULL,
  n_align = 0,
  n_align_tag = "NH",
  event_filters = c(0, 0, 0, 0, 0, 0, 0),
  only_keep_variants = FALSE,
  reads = NULL,
  return_data = TRUE
)

Arguments

bamfile

paths to 1 or 2 bam files, pass 2 files to

fafile

path to fasta file

bedfile

path to bed file with sites or regions to query

region

samtools region query string (i.e. chr1:100-1000)

chrom

chrom to process, not to be used with region

start

start position, not to be used with region

end

end position, not to be used with region

min_reads

min read depth needed to report site

min_base_qual

min base quality score to consider read for pileup

min_mapq

minimum required MAPQ score

max_depth

maximum read depth considered at each site

library_type

read orientation, one of fr-first-strand, fr-second-strand, or unstranded.

outfile

output tabix indexed file?

bedidx

A BedFile object, if supplied, pileup will use index generated by indexBed()

bam_flags

bam flags to filter or keep, use Rsamtools::scanBamFlag() to set

n_align

The number of valid alignments to use for each read, based on tag value supplied to n_align_tag. NULL disables filtering by n_align_tag value

n_align_tag

Tag name that containing # of valid alignments

event_filters

integer vector of length 4 with values indicating distance from pileup position to features to filter. Distance to trim from 5' ends of reads, from 3' end of reads, distance from splicing events, distance from indels, and length of homopolymer stretch used to determine if a site is in a homopolymer. The number of read mismatch types (e.g. A->G or G->T) allowed in a single read, and the required # of any mismatches that activate the mismatch type filter. default is no filters applied (e.g. c(0,0,0,0,0,0,0)).

only_keep_variants

if TRUE, then only variant sites will be reported (FALSE by default)

reads

if supplied a fasta file will be written with reads that pass filters and contain variants

return_data

return data as a Granges table?

Details

To compare RNA-seq to DNA-seq 2 bam files can be supplied to the bamfile argument, with the first corresponding to RNA-seq and second DNA-seq. The min_reads and min_mapq arguments can accept 2 values (as numeric vector) corresponding to filters for RNA and DNA-seq respectively. The min_base_qual, n_align, n_align_tag, and event_filters params will be applied to both files.

The library type argument will only apply to the 1st bam file. The second will be set to "unstranded"