Extract regions surrounding splice sites

Extract intervals at splice sites and their adjacent regions.

Usage

get_splice_sites(txdb, slop = 4)

Arguments

txdb

GenomicFeatures::TxDb

slop

The number of bases upstream and downstream of splice site to extract

Value

GenomicRanges::GRanges containing positions of splice sites, with flanking bases.

Examples

if (require(TxDb.Hsapiens.UCSC.hg38.knownGene)) {
    txdb <- TxDb.Hsapiens.UCSC.hg38.knownGene
    res <- get_splice_sites(txdb)
    res[1:5]
}
#> Loading required package: TxDb.Hsapiens.UCSC.hg38.knownGene
#> GRanges object with 5 ranges and 0 metadata columns:
#>       seqnames      ranges strand
#>          <Rle>   <IRanges>  <Rle>
#>   [1]     chr1 12054-12061      +
#>   [2]     chr1 12175-12182      +
#>   [3]     chr1 12224-12231      +
#>   [4]     chr1 12224-12231      +
#>   [5]     chr1 12609-12616      +
#>   -------
#>   seqinfo: 711 sequences from an unspecified genome; no seqlengths