Generate a samtools style index for a BedFile

Generate a samtools style index for a BedFile

Usage

indexBed(file, ...)

# S4 method for character
indexBed(file)

Arguments

file

path to bed file

...

For the generic, further arguments to pass to specific methods. Unused for now.

Value

Return an object of class BedFile which contains a pointer to the bed file index.

Examples

bed_fn <- system.file("extdata", "regions.bed", package = "raer")
indexBed(bed_fn)
#> Reference class object of class "BedFile"
#> Field ".extptr":
#> Object of class 'externalptr':
#>   x address: 0x559bcb3b0ba8
#>   R_ExternalPtrAddr(x): 0x559bb5c55420
#>   R_ExternalPtrTag(x): 0x559b9b4ef930
#>   typeof(R_ExternalPtrTag(x)): NULL
#>   R_ExternalPtrProtected(x): 0x559bd844c570
#>   typeof(R_ExternalPtrProtected(x)): character
#> Field "path":
#> [1] "/home/runner/work/_temp/Library/raer/extdata/regions.bed"
#> Field "open":
#> [1] TRUE