Generate base counts using pileup

Generate base counts using pileup

Usage

pileup_sites(
  bamfiles,
  fafile,
  bedfile = NULL,
  region = NULL,
  chroms = NULL,
  param = FilterParam(),
  outfile_prefix = NULL,
  reads = NULL,
  return_data = TRUE,
  BPPARAM = SerialParam(),
  use_index = FALSE,
  bad_reads = NULL,
  umi_tag = NULL,
  verbose = FALSE
)

FilterParam(
  max_depth = 10000,
  min_depth = 1L,
  min_base_quality = 20L,
  min_mapq = 0L,
  library_type = "fr-first-strand",
  bam_flags = NULL,
  only_keep_variants = FALSE,
  trim_5p = 0L,
  trim_3p = 0L,
  ftrim_5p = 0,
  ftrim_3p = 0,
  indel_dist = 0L,
  splice_dist = 0L,
  min_splice_overhang = 0L,
  homopolymer_len = 0L,
  max_mismatch_type = c(0L, 0L),
  read_bqual = c(0, 0),
  min_variant_reads = 0L,
  min_allelic_freq = 0,
  report_multiallelic = TRUE,
  ignore_query_Ns = FALSE
)

Arguments

bamfiles

character vector of paths to 1 or more bam files. If named, the names will be included in the colData of the RangedSummarizedExperiment, otherwise the colData will be populated with the basename of the bamfile.

fafile

path to fasta file

bedfile

path to bed file with sites or regions to query

region

samtools region query string (i.e. chr1:100-1000)

chroms

chromosomes to process, not to be used with region

param

object of class FilterParam() which specify various filters to apply to reads and sites during pileup.

outfile_prefix

Output prefix for tabix indexed files. If NULL, no files will be produced.

reads

if supplied a fasta file will be written with reads that pass filters and contain variants

return_data

if TRUE, data is returned as a RangedSummarizedExperiment, if FALSE a character vector of tabix-index files, specified by outfile_prefix, will be returned.

BPPARAM

A BiocParallel class to control parallel execution. Parallel processing occurs per chromosome, so is disabled when run on a single region.

use_index

if TRUE regions supplied in the bedfile will be queried using the bam file index. By default (FALSE) all alignments are queried via streaming and regions in the bedfile are used to select the sites to keep. Using the index can be much faster when querying a small number of sites in large bamfiles. With many sites in the bedfile , or with small bamfiles, this option may be slower than streaming.

bad_reads

a textfile containing read names to exclude from pileup. Readnames should be formated as readid_1 or readid_2 or readid for paired end first read paired-end second read or single end data.

umi_tag

The bam tag containing a UMI sequence. If supplied, multiple reads with the same UMI sequence will only be counted once per position.

verbose

if TRUE, then report progress and warnings.

max_depth

maximum read depth considered at each site

min_depth

min read depth needed to report site

min_base_quality

min base quality score to consider read for pileup

min_mapq

minimum required MAPQ score, can be a vector of values for each bam file

library_type

read orientation, one of fr-first-strand, fr-second-strand, unstranded, and genomic-unstranded. Can supply as a vector to specify for each input bam. Unstranded library type will be reported based on read alignment. genomic-unstranded will report all variants w.r.t the + strand.

bam_flags

bam flags to filter or keep, use Rsamtools::scanBamFlag() to generate.

only_keep_variants

if TRUE, then only variant sites will be reported (FALSE by default), can be a vector for each input bamfile

trim_5p

Bases to trim from 5' end of read alignments

trim_3p

Bases to trim from 3' end of read alignments

ftrim_5p

Fraction of bases to trim from 5' end of read alignments

ftrim_3p

Fraction of bases to trim from 3' end of read alignments

indel_dist

Exclude read if site occurs within given distance from indel event in the read

splice_dist

Exclude read if site occurs within given distance from splicing event in the read

min_splice_overhang

Exclude read if site is located adjacent to splice site with an overhang of less than given length.

homopolymer_len

Exclude site if occurs within homopolymer of given length

max_mismatch_type

Exclude read if it has X different mismatch types (e.g A-to-G, G-to-C, C-to-G, is 3 mismatch types) or Y # of mismatches, must be supplied as a integer vector of length 2. e.g. c(X, Y).

read_bqual

Exclude read if more than X percent of the bases have base qualities less than Y. Numeric vector of length 2. e.g. c(0.25, 20)

min_variant_reads

Required number of reads containing a variant for a site to be reported. Calculated per bam file, such that if 1 bam file has >= min_variant_reads, then the site will be reported.

min_allelic_freq

minimum allelic frequency required for a variant to be reported in Var assays.

report_multiallelic

if TRUE, report sites with multiple variants passing filters. If FALSE, site will not be reported.

ignore_query_Ns

ignored for now

Value

A RangedSummarizedExperiment or a vector of the output tabixed file names if return_data is FALSE.

Details

Multiple bam files can be processed together, with files being written for each bam file. In this mode the output regions will be consistent across all files. The min_mapq, only_keep_variants, and library_type parameters can be specified for each input files.

See also

Examples

library(SummarizedExperiment)
bamfn <- raer_example("SRR5564269_Aligned.sortedByCoord.out.md.bam")
bam2fn <- raer_example("SRR5564277_Aligned.sortedByCoord.out.md.bam")
fafn <- raer_example("human.fasta")

rse <- pileup_sites(bamfn, fafn)

fp <- FilterParam(only_keep_variants = TRUE, min_depth = 55)
pileup_sites(bamfn, fafn, param = fp)
#> class: RangedSummarizedExperiment 
#> dim: 7 1 
#> metadata(0):
#> assays(7): Var nRef ... nC nG
#> rownames(7): DHFR_207_- DHFR_208_- ... DHFR_300_- DHFR_332_-
#> rowData names(3): Ref rbpz vpb
#> colnames(1): SRR5564269_Aligned.sortedByCoord.out.md.bam
#> colData names(1): sample

# using multiple bam files

bamfn <- raer_example("SRR5564269_Aligned.sortedByCoord.out.md.bam")
bam2fn <- raer_example("SRR5564277_Aligned.sortedByCoord.out.md.bam")
fafn <- raer_example("human.fasta")

bams <- rep(c(bamfn, bam2fn), each = 3)
sample_ids <- paste0(rep(c("KO", "WT"), each = 3), 1:3)
names(bams) <- sample_ids

fp <- FilterParam(only_keep_variants = TRUE)
rse <- pileup_sites(bams, fafn, param = fp)
rse
#> class: RangedSummarizedExperiment 
#> dim: 74 6 
#> metadata(0):
#> assays(7): Var nRef ... nC nG
#> rownames(74): SSR3_102_- SSR3_125_- ... DHFR_430_- DHFR_513_-
#> rowData names(3): Ref rbpz vpb
#> colnames(6): KO1 KO2 ... WT2 WT3
#> colData names(1): sample

rse$condition <- substr(rse$sample, 1, 2)
assays(rse)
#> List of length 7
#> names(7): Var nRef nVar nA nT nC nG

colData(rse)
#> DataFrame with 6 rows and 2 columns
#>          sample   condition
#>     <character> <character>
#> KO1         KO1          KO
#> KO2         KO2          KO
#> KO3         KO3          KO
#> WT1         WT1          WT
#> WT2         WT2          WT
#> WT3         WT3          WT

rowRanges(rse)
#> GRanges object with 74 ranges and 3 metadata columns:
#>              seqnames    ranges strand |         Ref       rbpz
#>                 <Rle> <IRanges>  <Rle> | <character>  <numeric>
#>   SSR3_102_-     SSR3       102      - |           T   2.611800
#>   SSR3_125_-     SSR3       125      - |           C   0.623238
#>   SSR3_156_-     SSR3       156      - |           C   1.875144
#>   SSR3_176_-     SSR3       176      - |           A  -0.676423
#>   SSR3_198_-     SSR3       198      - |           A   1.817678
#>          ...      ...       ...    ... .         ...        ...
#>   DHFR_397_-     DHFR       397      - |           A -2.7341600
#>   DHFR_399_-     DHFR       399      - |           G -0.2094944
#>   DHFR_423_-     DHFR       423      - |           T -0.0815516
#>   DHFR_430_-     DHFR       430      - |           A -2.6781858
#>   DHFR_513_-     DHFR       513      - |           T -1.2475213
#>                      vpb
#>                <numeric>
#>   SSR3_102_- 2.21621e-02
#>   SSR3_125_- 2.21621e-02
#>   SSR3_156_- 2.21621e-02
#>   SSR3_176_- 2.00523e-05
#>   SSR3_198_- 2.21621e-02
#>          ...         ...
#>   DHFR_397_-   0.0221621
#>   DHFR_399_-   0.0221621
#>   DHFR_423_-   0.0221621
#>   DHFR_430_-   0.0221621
#>   DHFR_513_-   0.0221621
#>   -------
#>   seqinfo: 3 sequences from an unspecified genome