Calculate editing frequencies per cell or per cluster

Calculate editing frequencies per cell or per cluster

Usage

sc_editing(
  bamfile,
  fafile,
  bedfile,
  cell_barcodes,
  assay_cols = c("nA", "nG"),
  tag_index_args = list(tag = "CB"),
  BPPARAM = SerialParam(),
  batch_size = 50,
  umi_tag = "UB",
  param = FilterParam(),
  verbose = TRUE,
  ...
)

Arguments

bamfile

BAM file name

fafile

FASTA file name

bedfile

BED file containing editing sites

cell_barcodes

A list of character vectors containing cell barcodes to query or a single character vector of single cells to process. If a list is supplied it is assumed that alignments from cell barcodes in each list element should be pooled. If a single character vector, it is assumed that each cell barcode should be processed independently. See examples for specification.

assay_cols

assays to store in returned se. Set to "A" and "G". Note that storing multiple assays can require large amounts of memory.

tag_index_args

arguments pass to build_tag_index()

BPPARAM

BiocParallel instance. Parallel computation occurs across each entry in the cell_barcodes list, or across batches of single cells specified by batch_size.

batch_size

When processing single cells, the batch_size controls how many individual cell bams to process in each invocation of pileup_sites(). Batching the cells reduces run time by avoiding loading sequences from the fasta file for each cell. Setting values above 50 is unlikely to further improve runtime.

umi_tag

See arguments for [pileup_sites()]

param

object of class FilterParam() which specify various filters to apply to reads and sites during pileup.

verbose

Display messages

...

additional arguments passed to [pileup_sites()].

Examples

suppressPackageStartupMessages(library(SummarizedExperiment))

# get vector of cell barcodes in bam file (for use in this example)
# usually these would come from the single cell analysis

bamfn <- raer_example("5k_neuron_mouse_xf25_1pct_cbsort.bam")
idxfn <- build_tag_index(bamfn)
cbs <- show_tag_index(bamfn)$tag
cbs[1:5]
#> [1] "AAACCCAGTCACTTAG-1" "AAACGAACAACTGAAA-1" "AAACGAAGTGAATATG-1"
#> [4] "AAACGCTAGACTTGTC-1" "AAACGCTGTGGATTTC-1"

# process each cell individually
# will be slow with many sites and cells
# bam file will be indexed by build_tag_index() if not already done.
fp <- FilterParam(library_type = "fr-second-strand")
se <- sc_editing(
  bamfile = bamfn,
  fafile = raer_example("mouse_tiny.fasta"),
  bedfile = raer_example("5k_neuron_sites.bed.gz"),
  cell_barcodes = cbs[1:15],
  verbose = FALSE,
  param = fp
)

# pool cell barcodes across clusters
# pass a named list, with each list entry corresponding to a vector
# of cell barcodes from a cluster

# simulate 5 clusters
cb_lst <- split(cbs, cut(seq_along(cbs), breaks = 5))
names(cb_lst) <- paste0("cluster", 1:5)

se <- sc_editing(
  bamfile = bamfn,
  fafile = raer_example("mouse_tiny.fasta"),
  bedfile = raer_example("5k_neuron_sites.bed.gz"),
  cell_barcodes = cb_lst,
  umi_tag = NULL,
  verbose = FALSE,
  param = fp
)
assays(se)$nA
#> 17 x 5 sparse Matrix of class "dgCMatrix"
#>          cluster1 cluster2 cluster3 cluster4 cluster5
#> 2_548_-         1        2        1        3        2
#> 2_576_-        60       58       56       55       63
#> 2_591_-        94       92       99       96      105
#> 2_593_-        99       95      105      104      108
#> 2_609_-       161      152      162      168      159
#> 6_271_-         6        9        7        5       10
#> 6_331_-        34       40       34       34       54
#> 6_352_-        32       34       30       29       49
#> 6_361_-        23       26       25       24       39
#> 6_383_-        15       22       15       18       26
#> 6_393_-        13       15        8       10       12
#> 11_366_-        .        .        .        .        .
#> 11_367_-        .        .        .        .        .
#> 2_754_-         .        .        1        .        2
#> 2_755_-         .        .        1        .        2
#> 2_756_-         .        .        1        .        2
#> 2_792_-         .        .        .        .        2
assays(se)$nG
#> 17 x 5 sparse Matrix of class "dgCMatrix"
#>          cluster1 cluster2 cluster3 cluster4 cluster5
#> 2_548_-         .        .        1        .        .
#> 2_576_-         1        .        .        1        .
#> 2_591_-         .        1        1        1        .
#> 2_593_-         .        1        .        .        .
#> 2_609_-         1        .        .        .        .
#> 6_271_-         .        1        .        .        .
#> 6_331_-         1        1        .        .        .
#> 6_352_-         .        .        .        2        .
#> 6_361_-         4        2        .        2        2
#> 6_383_-         1        .        .        .        .
#> 6_393_-         1        .        .        .        .
#> 11_366_-        .        1        .        .        .
#> 11_367_-        .        1        .        .        .
#> 2_754_-         .        .        .        .        1
#> 2_755_-         .        .        .        .        1
#> 2_756_-         .        .        .        .        1
#> 2_792_-         .        .        1        .        1